+++ .eslintrc.cjs
@@ -1,18 +1,24 @@
 module.exports = {
-  env: {
-    node: true,
-  },
-  root: true,
-  parser: "@typescript-eslint/parser",
-  extends: [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/recommended",
-    "prettier",
-  ],
-  plugins: ["@typescript-eslint", "eslint-plugin-prettier"],
-  rules: {
-    "no-mixed-spaces-and-tabs": ["error", "smart-tabs"],
-    "no-unused-vars": "off",
-    "sort-imports": "off",
-  },
-};
+	env: {
+		browser: true,
+	},
+	globals: {
+		request: true,
+	},
+	parserOptions: {
+		ecmaVersion: 'latest',
+		sourceType: 'module',
+	},
+	extends: [
+		'eslint:recommended',
+		'plugin:@typescript-eslint/recommended',
+		'plugin:vue/vue3-recommended',
+		'@vue/typescript/recommended',
+		'prettier',
+	],
+	rules: {
+		'no-mixed-spaces-and-tabs': ['error', 'smart-tabs'],
+		'no-unused-vars': 'off',
+		'sort-imports': 'off',
+	},
+}