+++ vite.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig } from 'vitest/config'
+import Vue from '@vitejs/plugin-vue'
 import ESLint from 'vite-plugin-eslint'
 import path from 'path'
 
@@ -33,6 +34,14 @@ export default () => {
 				},
 			},
 		},
-		plugins: [ESLint()],
+		plugins: [
+			// https://github.com/vitejs/vite-plugin-vue
+			Vue({
+				include: [/\.vue$/],
+			}),
+
+			// https://github.com/gxmari007/vite-plugin-eslint
+			ESLint({ exclude: ['**/node_modules/**'] }),
+		],
 	})
 }