+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.2] - 2023-04-20
+
+### Added
+
+- `ReadProvider`, `SubmitProvider` and `RemoveProvider` components;
+- `immediate` option to `remove()` action.
+
 ## [1.0.1] - 2023-04-19
 
 ### Fix
@@ -46,6 +53,7 @@ All notable changes to this project will be documented in this file.
 - `read` and `submit` actions;
 - `getQueryByName` and `getItemByKey` getters.
 
+[1.0.2]: https://github.com/volverjs/query-vue/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/volverjs/query-vue/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/volverjs/query-vue/compare/v0.0.3...v1.0.0
 [0.0.3]: https://github.com/volverjs/query-vue/compare/v0.0.2...v0.0.3